+++ src/pages/pills-list/pills-list.ts
@@ -3,6 +3,7 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import { pacientesService } from '../../services/pacientes.service';
 import { AddPastillaPage } from '../add-pastilla/add-pastilla';
 import { SelectorPage } from '../selector/selector';
+import { AlertController } from 'ionic-angular';
 
 /**
  * Generated class for the PillsListPage page.
@@ -19,7 +20,7 @@ import { SelectorPage } from '../selector/selector';
 export class PillsListPage {
   item:any;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, public PacientesService : pacientesService) {
+  constructor(public navCtrl: NavController, public navParams: NavParams, public PacientesService : pacientesService, public alertCtrl: AlertController) {
     this.item = navParams.get('item');
    
   }
@@ -29,7 +30,27 @@ export class PillsListPage {
   }
 
   deletePastilla(paciente,pastilla){
-    this.PacientesService.deletePastilla(paciente,pastilla);
+    const confirm = this.alertCtrl.create({
+      title: 'Borrar Pastilla',
+      message: 'Está seguro que desea eliminar esta pastilla?',
+      buttons: [
+        {
+          text: 'Cancelar',
+          handler: () => {
+            console.log('Disagree clicked');
+          }
+        },
+        {
+          text: 'Borrar',
+          handler: () => {
+            this.PacientesService.deletePastilla(paciente,pastilla);
+            console.log('Agree clicked');
+          }
+        }
+      ]
+    });
+    confirm.present();
+    
   }
 
   goToAddPage(item){
